Bug: navbar looked blurry on chrome when zoomed in
Only happens when you're in the navbar--compact state. This was because
we left it with a transform3d still applied, which fuzzes the text a bit
(probably doesn't want to run nice text anti-aliasing for perf reasons).

Fix is to use keyframe animations and set top when the animation is
done.