Get project tiles in better shape
1) Give the space between them tiles equal space for top and bottom
(.col-equalspace-md-up adds that)
2) Make them clickable
3) Style them with dropshadows

For the shadows, also experimented with actually scaling the tile up
itself. This has a few issues: the text looks bad when scaled, weird
jitters if you use scale3d(), transformZ() also yields weird jitters in
the navbar.

We'll settle for this and use the illusion of scaling by zooming in on
the img a bit.