Responsive images
Use <picture> or srcset for responsive images. Calculated what image
widths are most important to us, ended up with a set of 6. Decided not
to use the picturefill polyfill. We do set <img src=""> as a fallback.
The only browsers that don't support <picture> at this time are IE 11,
some older versions of desktop Safari, and older versions of iOS Safari.
Since we serve a good resolution for iOS safari, they'll be in good
shape. They won't get the art direction, but that's not a huge deal.
Desktop Safari users will still work, and there's not many of them. IE
11 users shouldn't know the difference because most IE 11 machines are
not hooked up to high DPI displays.

Getting the scale animation to be smooth across browsers was tricky. You
can't have backface-visiblity: hidden, or IE11 will make the image semi
transparent while animating. You also can't scale by 1.011, or Firefox
will put in weird anti-ailiasing jitters. Everybody seems to be happy
with 1.012.

Also moved the scale selector to under .project-tile so that it wouldn't
un-scale the image when you hovered over the .project-tile__label.