Polish for project-tile on smaller screens
Initially explored having full-bleed width and height images for phone.
The background-image code is for that. Committing it just to record it,
I don't think it's a good option anymore because presses (with the tile
shrink effect) feel really wrong when it's full bleed.

The simple answer is to change the aspect ratio of the image to be more
at home on a phone.

Deleted the full-bleed grid classes since we aren't using them anymore.
Maybe they'll be brought back for blog posts.