Prevent FOUT with Typekit fonts
http://blog.typekit.com/2011/05/25/loading-typekit-fonts-asynchronously/
was helpful. The advanced embed code first adds .wf-loading to the html
element, then either removes it if fonts load within 3 seconds or adds
.wf-inactive if fonts failed to load. We only care about .wf-loading to
hide the fonts.

Had to surgically hide instead of just hiding large sets of tags because
an <a> isn't always used the same by everyone, etc.