Feature img loading on post.hbs and centered icon features
1) Enabled fade in loading for the feature image of post.hbs. This
actually fixed a bug where a low-res feature image would occasionally
get selected. The bug was a race condition: the code in main.js runs
after image downloading has been started, but does not wait for the
images to finish loading. The file is cached, so it was possible for the
main.js code to run *before* <img>'s srcset selected an image. Thus, it
went with whatever was on the src attribute, which is a low-res image.
The fix is to wait for the image to load, so zImgLoadedSwitchToBgImg was
added to <head> and it's called by an inline img.onload. The fade-in
required a new <div> to be added to post.hbs since
zImgLoadedSwitchToBgImg needs another element to hand the background
image off to because the <img> gets hidden (I tried to see if it could
all be done with a single <img> by clearning the src after we had a
correct image, but that leads to the "broken image" icon showing on
Firefox, which isn't something you can make transparent).

2) The /word-on-tablet/ and /cortana/ sections look goofy if their icons
aren't centered in the feature space. The ONLY way to do this is to set
the background to a fixed width. Trust me, I tried playing with
background-size: 100% auto, and background-position: center some% or
background-size: cover/contain. NONE of it works. Why? Because when you
set background-attachment: fixed, the coordinate system is for the
entire browser viewport, not the element. Thus, the height of the
browser, not the element, will affect where the image shows. The only
way to deal with this is to fix it in space. Luckily this works because
we can set a background-color that matches the image's background.

In short, the YAML isFeatureImgIconSize property is a super big hack,
and will only be used by these two /work/ pages.