Dial back the styling of my name
I'm not good enough at logo design to pull off the W in a box, plus it looks off vertical center when the top navigation is fixed to the page while you're scrolling. That's because the all caps links on the right are vertically centered to their cap height, which looks great. But then my name isn't all caps, and it's taller, so it looks too high. But your eye still wants them to be base-aligned. Tough problem.