Upgrade packages
We get security warnings due to gulp and gulp-imagemin dependencies. Unfortunately, upgrading to the next version (gulp 4) has a lot of breaking changes that will require rewriting the gulpfile.

Looked into the security issues, they're only a risk for client-facing code. Since this is just a static-site gen, we're fine.